# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/dental-lab/disputes/{id}": {
      "get": {
        "operationId": "DentalLabDisputesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-disputes"]
      }
    },
    "/api/dental-lab/disputes/{id}/messages": {
      "post": {
        "operationId": "DentalLabDisputesController_postMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-disputes"]
      }
    },
    "/api/dental-lab/disputes/{id}/close": {
      "post": {
        "operationId": "DentalLabDisputesController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDisputeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-disputes"]
      }
    },
...
    "/api/clinic/lab-disputes": {
      "get": {
        "operationId": "ClinicLabDisputesController_list",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-disputes"]
      },
      "post": {
        "operationId": "ClinicLabDisputesController_open",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDisputeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-disputes"]
      }
    },
    "/api/clinic/lab-disputes/{id}": {
      "get": {
        "operationId": "ClinicLabDisputesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-disputes"]
      }
    },
    "/api/clinic/lab-disputes/{id}/messages": {
      "post": {
        "operationId": "ClinicLabDisputesController_postMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-disputes"]
      }
    },
    "/api/clinic/lab-disputes/{id}/close": {
      "post": {
        "operationId": "ClinicLabDisputesController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDisputeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-disputes"]
      }
    },
...
  },
  "info": {
    "title": "ClinIQ API",
    "description": "Multi-tenant clinic management API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4005"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DentalLabCaseDisputeKind": {
        "type": "string",
        "enum": ["QUALITY", "BILLING", "DELIVERY", "OTHER"]
      },
      "OpenDisputeDto": {
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "description": "Case the dispute attaches to."
          },
          "kind": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DentalLabCaseDisputeKind"
              }
            ]
          },
          "reason": {
            "type": "string",
            "description": "Why the dispute is being opened."
          }
        },
        "required": ["caseId", "kind", "reason"]
      },
      "DisputeMessageDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        },
        "required": ["body"]
      },
      "DentalLabCaseDisputeStatus": {
        "type": "string",
        "enum": ["RESOLVED", "REJECTED", "WITHDRAWN"]
      },
      "ResolveDisputeDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DentalLabCaseDisputeStatus"
              }
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": ["status"]
      },
...
    }
  }
}